BackEnd/api_gateway/app/main.py
from fastapi import FastAPI, Request, Depends
from slowapi import Limiter, _rate_limit_exceeded_handler
from slowapi.util import get_remote_address
from slowapi.errors import RateLimitExceeded
from prometheus_fastapi_instrumentator import Instrumentator
from opentelemetry.instrumentation.fastapi import FastAPIInstrumentor
from backend.shared.auth import get_current_user
import structlog

logger = structlog.get_logger()

app = FastAPI(title="AskCart API Gateway")
limiter = Limiter(key_func=get_remote_address)
app.state.limiter = limiter
app.add_exception_handler(RateLimitExceeded, _rate_limit_exceeded_handler)

Instrumentator().instrument(app).expose(app)
FastAPIInstrumentor.instrument_app(app)

@app.middleware("http")
async def log_requests(request: Request, call_next):
    logger.info("Request", method=request.method, url=str(request.url))
    response = await call_next(request)
    return response

import httpx

PRODUCT_SERVICE_URL = "http://product-service:8000"

@app.get("/products/{product_id}")
@limiter.limit("5/minute")
async def get_product(product_id: int, user: str = Depends(get_current_user)):
    async with httpx.AsyncClient() as client:
        response = await client.get(f"{PRODUCT_SERVICE_URL}/products/{product_id}")
        if response.status_code != 200:
            raise HTTPException(status_code=response.status_code, detail=response.text)
        return response.json()

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)