requirements-base.txt
fastapi==0.115.0
uvicorn==0.31.0
pydantic==2.9.2
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
redis[asyncio]==5.1.1
confluent-kafka==2.6.0
httpx==0.27.2
pyjwt==2.9.0
slowapi==0.1.9
structlog==24.4.0
prometheus-fastapi-instrumentator==7.0.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
pytest==8.3.3
pytest-asyncio==0.24.0